requirements.txt
aioquic==1.2.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
Cython==3.1.3
Flask==3.1.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==3.0.2
mitmproxy==12.1.2
mitmproxy-windows==0.12.7
mitmproxy_rs==0.12.7
msgpack==1.1.1
passlib==1.7.4
protobuf==5.29.3
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydivert==2.1.0
pylsqpack==0.3.22
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
service-identity==24.2.0
setuptools==80.9.0
sortedcontainers==2.4.0
tornado==6.5.2
typing_extensions==4.14.1
urwid==2.6.16
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.24.0
aiohttp==3.9.5
pycryptodome==3.18.0
requests==2.31.0
pyjwt